pyproject.toml
[project]
name = "lmbench"
version = "0.1.0"
description = "vLLM benchmark harness with NVFP4 quantization comparison."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Kai" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "plotly>=5.22",
    "httpx>=0.27",
    "tenacity>=8.3",
    "jinja2>=3.1",
]

[project.optional-dependencies]
gpu = [
    "vllm>=0.7",
    "transformers>=4.45",
    "datasets>=2.20",
    "accelerate>=0.33",
    "lm-eval>=0.4.4",
    "pynvml>=11.5",
]
quant = [
    "nvidia-modelopt[all]>=0.19",
]
all = [
    "lmbench[gpu,quant]",
]

[project.scripts]
lmbench = "lmbench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lmbench"]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.6",
    "types-pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/lmbench"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "pynvml.*",
    "vllm.*",
    "modelopt.*",
    "lm_eval.*",
    "pandas.*",
    "plotly.*",
    "torch.*",
    "transformers.*",
    "datasets.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=lmbench",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "gpu: requires a CUDA GPU",
    "blackwell: requires NVIDIA Blackwell (B100/B200/B300)",
    "slow: long-running test",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["lmbench"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
